backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlmodel==0.0.24
python-dotenv==1.1.0
sqlalchemy-libsql==0.2.0
pydantic-settings==2.8.1
